+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add calculated data documents to Moldev Softmax Pro
 - Add multi-analyte-profiling BENCHLING/2024/01 schema
 - Parser for Nanodrop Eight
+- Add --regex argument to schema generation script
+
 ### Fixed
 - Perkin Elmer Envision: calculated data name now captures string to the left - rather than right - of the ‘=’ in the Formula cell.